# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.UserStatsRequest:
    properties:
      activity:
        type: integer
      age:
        type: integer
      calories_goal:
        type: integer
      height:
        type: integer
      protein_ratio:
        type: integer
      user_id:
        type: integer
      weight:
        type: integer
    type: object
  model.UserStatsResponse:
    properties:
      activity:
        type: integer
      age:
        type: integer
      calories_goal:
        type: integer
      height:
        type: integer
      id:
        type: integer
      protein_ratio:
        type: integer
      weight:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample server for a sport application.
  title: Sportin API
  version: "1.0"
paths:
...
  /userStats:
    get:
      consumes:
      - application/json
      description: Get all UserStats
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.UserStatsResponse'
            type: array
      summary: Get all UserStats
      tags:
      - UserStats
    post:
      consumes:
      - application/json
      description: Create a new UserStats
      parameters:
      - description: UserStats object that needs to be created
        in: body
        name: userStats
        required: true
        schema:
          $ref: '#/definitions/model.UserStatsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserStatsResponse'
      summary: Create a new UserStats
      tags:
      - UserStats
  /userStats/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a UserStats
      parameters:
      - description: UserStats ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a UserStats
      tags:
      - UserStats
    get:
      consumes:
      - application/json
      description: Get UserStats
      parameters:
      - description: UserStats ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserStatsResponse'
      summary: Get UserStats
      tags:
      - UserStats
    put:
      consumes:
      - application/json
      description: Update a UserStats
      parameters:
      - description: UserStats ID
        in: path
        name: id
        required: true
        type: integer
      - description: UserStats object that needs to be updated
        in: body
        name: userStats
        required: true
        schema:
          $ref: '#/definitions/model.UserStatsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserStatsResponse'
      summary: Update a UserStats
      tags:
      - UserStats
...
swagger: "2.0"